requirements.txt
discord.py
watchfiles